// src/types/state/inventory/inventoryActionTypes.ts
export enum ActionTypes {
    SET_PRODUCTS = 'SET_PRODUCTS',
    NEW_PRODUCT = 'NEW_PRODUCT',
    EDIT_PRODUCT = 'EDIT_PRODUCT',
    SET_COLOURS = 'SET_COLOURS',
    NEW_COLOUR = 'NEW_COLOUR',
    EDIT_COLOUR = 'EDIT_COLOUR',
    SET_CATEGORIES = 'SET_CATEGORIES',
    NEW_CATEGORY = 'NEW_CATEGORY',
    EDIT_CATEGORY = 'EDIT_CATEGORY',
    SET_BRANDS = 'SET_BRANDS',
    NEW_BRAND = 'NEW_BRAND',
    EDIT_BRAND = 'EDIT_BRAND',
    SET_SIZES = 'SET_SIZES',
    NEW_SIZE = 'NEW_SIZE',
    EDIT_SIZE = 'EDIT_SIZE'
};